Fix deadlock when adding consumer
